+++ src/plugin/parquet-writer-hl.cpp
@@ -5,6 +5,9 @@
 // scalars
 //     __sparquet_ncol
 //     __sparquet_rg_size
+//     __sparquet_chunkbytes
+//     __sparquet_progress
+//     __sparquet_check
 ST_retcode sf_hl_write_varlist(
     const char *fname,
     const char *fcols,
@@ -13,16 +16,18 @@ ST_retcode sf_hl_write_varlist(
     const int strbuffer)
 {
 
-    ST_double z;
+    ST_double z, progress;
     ST_retcode rc = 0, any_rc = 0;
     SPARQUET_CHAR(vstr, strbuffer);
 
     int64_t in1 = SF_in1();
     int64_t in2 = SF_in2();
     int64_t N = in2 - in1 + 1;
-    int64_t vtype, i, j, ncol = 1, rg_size = 16;
-    clock_t timer = clock();
+    int64_t vtype, i, j, ttot, tread, tevery, ncol = 1, rg_size = 16;
+    int64_t arraybytes = 0, chunkbytes = 1073741824;
     int64_t warn_extended = 0;
+    clock_t timer = clock();
+    clock_t stimer = clock();
 
     std::string line;
     std::ifstream fstream;
@@ -30,16 +35,22 @@ ST_retcode sf_hl_write_varlist(
     // Get column and type info from Stata
     // -----------------------------------
 
-    if ( (rc = sf_scalar_int("__sparquet_rg_size", 18, &rg_size)) ) any_rc = rc;
-    if ( (rc = sf_scalar_int("__sparquet_ncol",    15, &ncol))    ) any_rc = rc;
+    if ( (rc = sf_scalar_int("__sparquet_chunkbytes", 21, &chunkbytes)) ) any_rc = rc;
+    if ( (rc = sf_scalar_int("__sparquet_rg_size",    18, &rg_size))    ) any_rc = rc;
+    if ( (rc = sf_scalar_int("__sparquet_ncol",       15, &ncol))       ) any_rc = rc;
+    if ( (rc = sf_scalar_dbl("__sparquet_progress",   19, &progress))   ) any_rc = rc;
+    if ( (rc = sf_scalar_int("__sparquet_check",      16, &tevery))     ) any_rc = rc;
 
     sf_printf_debug(debug, "# columns: %ld\n", ncol);
 
+    tread = 0;
+    ttot = ncol * N;
     int64_t vtypes[ncol];
     std::string vnames[ncol];
 
     std::vector<std::shared_ptr<arrow::Field>> vfields(ncol);
     std::vector<std::shared_ptr<arrow::Array>> varrays(ncol);
+    std::vector<std::shared_ptr<arrow::ChunkedArray>> carrays(ncol);
 
     if ( (rc = sf_matrix_int("__sparquet_coltypes", 19, ncol, vtypes)) ) any_rc = rc;
 
@@ -77,9 +88,11 @@ ST_retcode sf_hl_write_varlist(
         timer = clock();
         for (j = 0; j < ncol; j++) {
             vtype = vtypes[j];
+            arraybytes = 0;
+            std::vector<std::shared_ptr<arrow::Array>> chunks;
             // sf_printf_debug(2, "col %ld: %ld (%s)\n", j, vtype, vnames[j].c_str());
-            // TODO: Boolean is only 0/1; keep? Or always int32?
             if ( vtype == -1 ) {
+                // TODO: Boolean is only 0/1; keep? Or always int32?
                 arrow::BooleanBuilder boolbuilder;
                 for (i = 0; i < N; i++) {
                     if ( (rc = SF_vdata(j + 1, i + in1, &z)) ) goto exit;
@@ -92,8 +105,30 @@ ST_retcode sf_hl_write_varlist(
                             ++warn_extended;
                         }
                     }
+                    arraybytes += sizeof(bool);
+                    if ( arraybytes > chunkbytes ) {
+                        PARQUET_THROW_NOT_OK(boolbuilder.Finish(&varrays[j]));
+                        chunks.push_back(std::move(varrays[j]));
+                        boolbuilder.Reset();
+                        arraybytes = 0;
+                    }
+                    if ( i % tevery == 0 ) {
+                        tread += tevery;
+                        sf_running_progress_write(
+                            &timer,
+                            &stimer,
+                            progress,
+                            j + 1, ncol,
+                            i + 1, N,
+                            100 * tread / ttot
+                        );
+                    }
                 }
-                PARQUET_THROW_NOT_OK(boolbuilder.Finish(&varrays[j]));
+                if ( arraybytes ) {
+                    PARQUET_THROW_NOT_OK(boolbuilder.Finish(&varrays[j]));
+                    chunks.push_back(std::move(varrays[j]));
+                }
+                carrays[j] = std::make_shared<arrow::ChunkedArray>(std::move(chunks));
                 vfields[j] = arrow::field(vnames[j].c_str(), arrow::boolean());
             }
             else if ( vtype == -2 ) {
@@ -109,8 +144,30 @@ ST_retcode sf_hl_write_varlist(
                             ++warn_extended;
                         }
                     }
+                    arraybytes += sizeof(int32_t);
+                    if ( arraybytes > chunkbytes ) {
+                        PARQUET_THROW_NOT_OK(i32builder.Finish(&varrays[j]));
+                        chunks.push_back(std::move(varrays[j]));
+                        i32builder.Reset();
+                        arraybytes = 0;
+                    }
+                    if ( i % tevery == 0 ) {
+                        tread += tevery;
+                        sf_running_progress_write(
+                            &timer,
+                            &stimer,
+                            progress,
+                            j + 1, ncol,
+                            i + 1, N,
+                            100 * tread / ttot
+                        );
+                    }
                 }
-                PARQUET_THROW_NOT_OK(i32builder.Finish(&varrays[j]));
+                if ( arraybytes ) {
+                    PARQUET_THROW_NOT_OK(i32builder.Finish(&varrays[j]));
+                    chunks.push_back(std::move(varrays[j]));
+                }
+                carrays[j] = std::make_shared<arrow::ChunkedArray>(std::move(chunks));
                 vfields[j] = arrow::field(vnames[j].c_str(), arrow::int32());
             }
             else if ( vtype == -3 ) {
@@ -126,8 +183,30 @@ ST_retcode sf_hl_write_varlist(
                             ++warn_extended;
                         }
                     }
+                    arraybytes += sizeof(int32_t);
+                    if ( arraybytes > chunkbytes ) {
+                        PARQUET_THROW_NOT_OK(i32builder.Finish(&varrays[j]));
+                        chunks.push_back(std::move(varrays[j]));
+                        i32builder.Reset();
+                        arraybytes = 0;
+                    }
+                    if ( i % tevery == 0 ) {
+                        tread += tevery;
+                        sf_running_progress_write(
+                            &timer,
+                            &stimer,
+                            progress,
+                            j + 1, ncol,
+                            i + 1, N,
+                            100 * tread / ttot
+                        );
+                    }
                 }
-                PARQUET_THROW_NOT_OK(i32builder.Finish(&varrays[j]));
+                if ( arraybytes ) {
+                    PARQUET_THROW_NOT_OK(i32builder.Finish(&varrays[j]));
+                    chunks.push_back(std::move(varrays[j]));
+                }
+                carrays[j] = std::make_shared<arrow::ChunkedArray>(std::move(chunks));
                 vfields[j] = arrow::field(vnames[j].c_str(), arrow::int32());
             }
             else if ( vtype == -4 ) {
@@ -143,17 +222,38 @@ ST_retcode sf_hl_write_varlist(
                             ++warn_extended;
                         }
                     }
+                    arraybytes += sizeof(float);
+                    if ( arraybytes > chunkbytes ) {
+                        PARQUET_THROW_NOT_OK(f32builder.Finish(&varrays[j]));
+                        chunks.push_back(std::move(varrays[j]));
+                        f32builder.Reset();
+                        arraybytes = 0;
+                    }
+                    if ( i % tevery == 0 ) {
+                        tread += tevery;
+                        sf_running_progress_write(
+                            &timer,
+                            &stimer,
+                            progress,
+                            j + 1, ncol,
+                            i + 1, N,
+                            100 * tread / ttot
+                        );
+                    }
                 }
-                PARQUET_THROW_NOT_OK(f32builder.Finish(&varrays[j]));
+                if ( arraybytes ) {
+                    PARQUET_THROW_NOT_OK(f32builder.Finish(&varrays[j]));
+                    chunks.push_back(std::move(varrays[j]));
+                }
+                carrays[j] = std::make_shared<arrow::ChunkedArray>(std::move(chunks));
                 vfields[j] = arrow::field(vnames[j].c_str(), arrow::float32());
             }
             else if ( vtype == -5 ) {
-                arrow::FloatBuilder f32builder;
                 arrow::DoubleBuilder f64builder;
                 for (i = 0; i < N; i++) {
                     if ( (rc = SF_vdata(j + 1, i + in1, &z)) ) goto exit;
                     if ( z < SV_missval ) {
-                        PARQUET_THROW_NOT_OK(f64builder.Append((float) z));
+                        PARQUET_THROW_NOT_OK(f64builder.Append((double) z));
                     }
                     else {
                         PARQUET_THROW_NOT_OK(f64builder.AppendNull());
@@ -161,8 +261,30 @@ ST_retcode sf_hl_write_varlist(
                             ++warn_extended;
                         }
                     }
+                    arraybytes += sizeof(double);
+                    if ( arraybytes > chunkbytes ) {
+                        PARQUET_THROW_NOT_OK(f64builder.Finish(&varrays[j]));
+                        chunks.push_back(std::move(varrays[j]));
+                        f64builder.Reset();
+                        arraybytes = 0;
+                    }
+                    if ( i % tevery == 0 ) {
+                        tread += tevery;
+                        sf_running_progress_write(
+                            &timer,
+                            &stimer,
+                            progress,
+                            j + 1, ncol,
+                            i + 1, N,
+                            100 * tread / ttot
+                        );
+                    }
                 }
-                PARQUET_THROW_NOT_OK(f64builder.Finish(&varrays[j]));
+                if ( arraybytes ) {
+                    PARQUET_THROW_NOT_OK(f64builder.Finish(&varrays[j]));
+                    chunks.push_back(std::move(varrays[j]));
+                }
+                carrays[j] = std::make_shared<arrow::ChunkedArray>(std::move(chunks));
                 vfields[j] = arrow::field(vnames[j].c_str(), arrow::float64());
             }
             else if ( vtype > 0 ) {
@@ -170,9 +292,31 @@ ST_retcode sf_hl_write_varlist(
                 for (i = 0; i < N; i++) {
                     if ( (rc = SF_sdata(j + 1, i + in1, vstr)) ) goto exit;
                     PARQUET_THROW_NOT_OK(strbuilder.Append(vstr));
+                    arraybytes += std::strlen(vstr) * sizeof(vstr);
                     memset(vstr, '\0', strbuffer);
+                    if ( arraybytes > chunkbytes ) {
+                        PARQUET_THROW_NOT_OK(strbuilder.Finish(&varrays[j]));
+                        chunks.push_back(std::move(varrays[j]));
+                        strbuilder.Reset();
+                        arraybytes = 0;
+                    }
+                    if ( i % tevery == 0 ) {
+                        tread += tevery;
+                        sf_running_progress_write(
+                            &timer,
+                            &stimer,
+                            progress,
+                            j + 1, ncol,
+                            i + 1, N,
+                            100 * tread / ttot
+                        );
+                    }
                 }
-                PARQUET_THROW_NOT_OK(strbuilder.Finish(&varrays[j]));
+                if ( arraybytes ) {
+                    PARQUET_THROW_NOT_OK(strbuilder.Finish(&varrays[j]));
+                    chunks.push_back(std::move(varrays[j]));
+                }
+                carrays[j] = std::make_shared<arrow::ChunkedArray>(std::move(chunks));
                 vfields[j] = arrow::field(vnames[j].c_str(), arrow::utf8());
             }
             else {
@@ -184,7 +328,7 @@ ST_retcode sf_hl_write_varlist(
         sf_running_timer (&timer, "Copied data into Arrow table");
 
         std::shared_ptr<arrow::Schema> schema = arrow::schema(vfields);
-        std::shared_ptr<arrow::Table> table = arrow::Table::Make(schema, varrays);
+        std::shared_ptr<arrow::Table> table = arrow::Table::Make(schema, carrays);
 
         // rg_size is, according to the source files, supposed to be
         // really large and controls how the file gets split.
@@ -201,7 +345,7 @@ ST_retcode sf_hl_write_varlist(
         sf_printf_debug(verbose, "\t%ld rows\n",    N);
 
         if ( warn_extended > 0 ) {
-            sf_printf("Warning: %ld extended missing values coerced to NaN.\n", warn_extended);
+            sf_printf("Warning: %ld extended missing values coerced to NULL.\n", warn_extended);
         }
 
     } catch (const std::exception& e) {
@@ -220,6 +364,9 @@ exit:
 // scalars
 //     __sparquet_ncol
 //     __sparquet_rg_size
+//     __sparquet_chunkbytes
+//     __sparquet_progress
+//     __sparquet_check
 ST_retcode sf_hl_write_varlist_if(
     const char *fname,
     const char *fcols,
@@ -228,16 +375,18 @@ ST_retcode sf_hl_write_varlist_if(
     const int strbuffer)
 {
 
-    ST_double z;
+    ST_double z, progress;
     ST_retcode rc = 0, any_rc = 0;
     SPARQUET_CHAR(vstr, strbuffer);
 
     int64_t in1 = SF_in1();
     int64_t in2 = SF_in2();
     int64_t N = in2 - in1 + 1;
-    int64_t vtype, i, j, ncol = 1, rg_size = 16;
-    clock_t timer = clock();
+    int64_t vtype, i, j, ttot, tread, tevery, ncol = 1, rg_size = 16;
+    int64_t arraybytes = 0, chunkbytes = 1073741824;
     int64_t warn_extended = 0;
+    clock_t timer = clock();
+    clock_t stimer = clock();
 
     std::string line;
     std::ifstream fstream;
@@ -245,16 +394,22 @@ ST_retcode sf_hl_write_varlist_if(
     // Get column and type info from Stata
     // -----------------------------------
 
-    if ( (rc = sf_scalar_int("__sparquet_rg_size", 18, &rg_size)) ) any_rc = rc;
-    if ( (rc = sf_scalar_int("__sparquet_ncol",    15, &ncol))    ) any_rc = rc;
+    if ( (rc = sf_scalar_int("__sparquet_chunkbytes", 21, &chunkbytes)) ) any_rc = rc;
+    if ( (rc = sf_scalar_int("__sparquet_rg_size",    18, &rg_size))    ) any_rc = rc;
+    if ( (rc = sf_scalar_int("__sparquet_ncol",       15, &ncol))       ) any_rc = rc;
+    if ( (rc = sf_scalar_dbl("__sparquet_progress",   19, &progress))   ) any_rc = rc;
+    if ( (rc = sf_scalar_int("__sparquet_check",      16, &tevery))     ) any_rc = rc;
 
     sf_printf_debug(debug, "# columns: %ld\n", ncol);
 
+    tread = 0;
+    ttot = ncol * N;
     int64_t vtypes[ncol];
     std::string vnames[ncol];
 
     std::vector<std::shared_ptr<arrow::Field>> vfields(ncol);
     std::vector<std::shared_ptr<arrow::Array>> varrays(ncol);
+    std::vector<std::shared_ptr<arrow::ChunkedArray>> carrays(ncol);
 
     if ( (rc = sf_matrix_int("__sparquet_coltypes", 19, ncol, vtypes)) ) any_rc = rc;
 
@@ -292,9 +447,11 @@ ST_retcode sf_hl_write_varlist_if(
         timer = clock();
         for (j = 0; j < ncol; j++) {
             vtype = vtypes[j];
+            arraybytes = 0;
+            std::vector<std::shared_ptr<arrow::Array>> chunks;
             // sf_printf_debug(2, "col %ld: %ld (%s)\n", j, vtype, vnames[j].c_str());
-            // TODO: Boolean is only 0/1; keep? Or always int32?
             if ( vtype == -1 ) {
+                // TODO: Boolean is only 0/1; keep? Or always int32?
                 arrow::BooleanBuilder boolbuilder;
                 for (i = 0; i < N; i++) {
                     if ( SF_ifobs(i + in1) ) {
@@ -308,9 +465,31 @@ ST_retcode sf_hl_write_varlist_if(
                                 ++warn_extended;
                             }
                         }
+                        arraybytes += sizeof(bool);
+                        if ( arraybytes > chunkbytes ) {
+                            PARQUET_THROW_NOT_OK(boolbuilder.Finish(&varrays[j]));
+                            chunks.push_back(std::move(varrays[j]));
+                            boolbuilder.Reset();
+                            arraybytes = 0;
+                        }
+                    }
+                    if ( i % tevery == 0 ) {
+                        tread += tevery;
+                        sf_running_progress_write(
+                            &timer,
+                            &stimer,
+                            progress,
+                            j + 1, ncol,
+                            i + 1, N,
+                            100 * tread / ttot
+                        );
                     }
                 }
-                PARQUET_THROW_NOT_OK(boolbuilder.Finish(&varrays[j]));
+                if ( arraybytes ) {
+                    PARQUET_THROW_NOT_OK(boolbuilder.Finish(&varrays[j]));
+                    chunks.push_back(std::move(varrays[j]));
+                }
+                carrays[j] = std::make_shared<arrow::ChunkedArray>(std::move(chunks));
                 vfields[j] = arrow::field(vnames[j].c_str(), arrow::boolean());
             }
             else if ( vtype == -2 ) {
@@ -327,9 +506,31 @@ ST_retcode sf_hl_write_varlist_if(
                                 ++warn_extended;
                             }
                         }
+                        arraybytes += sizeof(int32_t);
+                        if ( arraybytes > chunkbytes ) {
+                            PARQUET_THROW_NOT_OK(i32builder.Finish(&varrays[j]));
+                            chunks.push_back(std::move(varrays[j]));
+                            i32builder.Reset();
+                            arraybytes = 0;
+                        }
+                    }
+                    if ( i % tevery == 0 ) {
+                        tread += tevery;
+                        sf_running_progress_write(
+                            &timer,
+                            &stimer,
+                            progress,
+                            j + 1, ncol,
+                            i + 1, N,
+                            100 * tread / ttot
+                        );
                     }
                 }
-                PARQUET_THROW_NOT_OK(i32builder.Finish(&varrays[j]));
+                if ( arraybytes ) {
+                    PARQUET_THROW_NOT_OK(i32builder.Finish(&varrays[j]));
+                    chunks.push_back(std::move(varrays[j]));
+                }
+                carrays[j] = std::make_shared<arrow::ChunkedArray>(std::move(chunks));
                 vfields[j] = arrow::field(vnames[j].c_str(), arrow::int32());
             }
             else if ( vtype == -3 ) {
@@ -346,9 +547,31 @@ ST_retcode sf_hl_write_varlist_if(
                                 ++warn_extended;
                             }
                         }
+                        arraybytes += sizeof(int32_t);
+                        if ( arraybytes > chunkbytes ) {
+                            PARQUET_THROW_NOT_OK(i32builder.Finish(&varrays[j]));
+                            chunks.push_back(std::move(varrays[j]));
+                            i32builder.Reset();
+                            arraybytes = 0;
+                        }
+                    }
+                    if ( i % tevery == 0 ) {
+                        tread += tevery;
+                        sf_running_progress_write(
+                            &timer,
+                            &stimer,
+                            progress,
+                            j + 1, ncol,
+                            i + 1, N,
+                            100 * tread / ttot
+                        );
                     }
                 }
-                PARQUET_THROW_NOT_OK(i32builder.Finish(&varrays[j]));
+                if ( arraybytes ) {
+                    PARQUET_THROW_NOT_OK(i32builder.Finish(&varrays[j]));
+                    chunks.push_back(std::move(varrays[j]));
+                }
+                carrays[j] = std::make_shared<arrow::ChunkedArray>(std::move(chunks));
                 vfields[j] = arrow::field(vnames[j].c_str(), arrow::int32());
             }
             else if ( vtype == -4 ) {
@@ -365,19 +588,40 @@ ST_retcode sf_hl_write_varlist_if(
                                 ++warn_extended;
                             }
                         }
+                        arraybytes += sizeof(float);
+                        if ( arraybytes > chunkbytes ) {
+                            PARQUET_THROW_NOT_OK(f32builder.Finish(&varrays[j]));
+                            chunks.push_back(std::move(varrays[j]));
+                            f32builder.Reset();
+                            arraybytes = 0;
+                        }
+                    }
+                    if ( i % tevery == 0 ) {
+                        tread += tevery;
+                        sf_running_progress_write(
+                            &timer,
+                            &stimer,
+                            progress,
+                            j + 1, ncol,
+                            i + 1, N,
+                            100 * tread / ttot
+                        );
                     }
                 }
-                PARQUET_THROW_NOT_OK(f32builder.Finish(&varrays[j]));
+                if ( arraybytes ) {
+                    PARQUET_THROW_NOT_OK(f32builder.Finish(&varrays[j]));
+                    chunks.push_back(std::move(varrays[j]));
+                }
+                carrays[j] = std::make_shared<arrow::ChunkedArray>(std::move(chunks));
                 vfields[j] = arrow::field(vnames[j].c_str(), arrow::float32());
             }
             else if ( vtype == -5 ) {
-                arrow::FloatBuilder f32builder;
                 arrow::DoubleBuilder f64builder;
                 for (i = 0; i < N; i++) {
                     if ( SF_ifobs(i + in1) ) {
                         if ( (rc = SF_vdata(j + 1, i + in1, &z)) ) goto exit;
                         if ( z < SV_missval ) {
-                            PARQUET_THROW_NOT_OK(f64builder.Append((float) z));
+                            PARQUET_THROW_NOT_OK(f64builder.Append((double) z));
                         }
                         else {
                             PARQUET_THROW_NOT_OK(f64builder.AppendNull());
@@ -385,9 +629,31 @@ ST_retcode sf_hl_write_varlist_if(
                                 ++warn_extended;
                             }
                         }
+                        arraybytes += sizeof(double);
+                        if ( arraybytes > chunkbytes ) {
+                            PARQUET_THROW_NOT_OK(f64builder.Finish(&varrays[j]));
+                            chunks.push_back(std::move(varrays[j]));
+                            f64builder.Reset();
+                            arraybytes = 0;
+                        }
+                    }
+                    if ( i % tevery == 0 ) {
+                        tread += tevery;
+                        sf_running_progress_write(
+                            &timer,
+                            &stimer,
+                            progress,
+                            j + 1, ncol,
+                            i + 1, N,
+                            100 * tread / ttot
+                        );
                     }
                 }
-                PARQUET_THROW_NOT_OK(f64builder.Finish(&varrays[j]));
+                if ( arraybytes ) {
+                    PARQUET_THROW_NOT_OK(f64builder.Finish(&varrays[j]));
+                    chunks.push_back(std::move(varrays[j]));
+                }
+                carrays[j] = std::make_shared<arrow::ChunkedArray>(std::move(chunks));
                 vfields[j] = arrow::field(vnames[j].c_str(), arrow::float64());
             }
             else if ( vtype > 0 ) {
@@ -396,10 +662,32 @@ ST_retcode sf_hl_write_varlist_if(
                     if ( SF_ifobs(i + in1) ) {
                         if ( (rc = SF_sdata(j + 1, i + in1, vstr)) ) goto exit;
                         PARQUET_THROW_NOT_OK(strbuilder.Append(vstr));
+                        arraybytes += std::strlen(vstr) * sizeof(vstr);
                         memset(vstr, '\0', strbuffer);
+                        if ( arraybytes > chunkbytes ) {
+                            PARQUET_THROW_NOT_OK(strbuilder.Finish(&varrays[j]));
+                            chunks.push_back(std::move(varrays[j]));
+                            strbuilder.Reset();
+                            arraybytes = 0;
+                        }
+                    }
+                    if ( i % tevery == 0 ) {
+                        tread += tevery;
+                        sf_running_progress_write(
+                            &timer,
+                            &stimer,
+                            progress,
+                            j + 1, ncol,
+                            i + 1, N,
+                            100 * tread / ttot
+                        );
                     }
                 }
-                PARQUET_THROW_NOT_OK(strbuilder.Finish(&varrays[j]));
+                if ( arraybytes ) {
+                    PARQUET_THROW_NOT_OK(strbuilder.Finish(&varrays[j]));
+                    chunks.push_back(std::move(varrays[j]));
+                }
+                carrays[j] = std::make_shared<arrow::ChunkedArray>(std::move(chunks));
                 vfields[j] = arrow::field(vnames[j].c_str(), arrow::utf8());
             }
             else {
@@ -411,7 +699,7 @@ ST_retcode sf_hl_write_varlist_if(
         sf_running_timer (&timer, "Copied data into Arrow table");
 
         std::shared_ptr<arrow::Schema> schema = arrow::schema(vfields);
-        std::shared_ptr<arrow::Table> table = arrow::Table::Make(schema, varrays);
+        std::shared_ptr<arrow::Table> table = arrow::Table::Make(schema, carrays);
 
         // rg_size is, according to the source files, supposed to be
         // really large and controls how the file gets split.
@@ -428,7 +716,7 @@ ST_retcode sf_hl_write_varlist_if(
         sf_printf_debug(verbose, "\t%ld rows\n",    N);
 
         if ( warn_extended > 0 ) {
-            sf_printf("Warning: %ld extended missing values coerced to NaN.\n", warn_extended);
+            sf_printf("Warning: %ld extended missing values coerced to NULL.\n", warn_extended);
         }
 
     } catch (const std::exception& e) {